chore: move views to examples
